Update 15/07
Nuevos ajustes. Para presentar

+++ src/main/java/demo/service/InstructorService.java
@@ -1,6 +1,7 @@
 package demo.service;
 
 import demo.model.Instructor;
+import demo.repository.HorarioRepository;
 import demo.repository.InstructorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -12,6 +13,9 @@ public class InstructorService {
 
     @Autowired
     private InstructorRepository instructorRepository;
+
+    @Autowired
+    private HorarioRepository horarioRepository;
 
     public List<Instructor> findAll() {
         return instructorRepository.findAll();
@@ -27,6 +31,10 @@ public class InstructorService {
     }
 
     public void deleteById(Long id) {
+        boolean tieneHorarios = horarioRepository.existsByInstructorId(id);
+        if (tieneHorarios) {
+            throw new RuntimeException("No se puede eliminar. El instructor tiene horarios asignados.");
+        }
         instructorRepository.deleteById(id);
     }
 }
